handled restructuring file/folder naming for admin folder
