+++ client/components/admin/AddEmployee/AddEmployee.css
@@ -1,0 +1,7 @@
+.add_employee {
+  display: block;
+  border: 1px solid black;
+  margin: 5px;
+  width: 40%;
+  border-radius: 10px;
+}
