+++ client/components/admin/AdminDashboard/AdminDashboard.css
@@ -1,0 +1,3 @@
+.admin-dashboard {
+  border: 3px solid rgb(7, 67, 7);
+}
