+++ client/components/admin/AdminDashboard/AdminDashboard.tsx
@@ -1,7 +1,9 @@
 import React, { Fragment } from 'react';
 import { Link } from 'react-router-dom';
-import EmployeeList from './EmployeeList'
-import AddEmployee from './AddEmployee';
+// These next two imports appear to not be needed.  Commenting them out to see if anything breaks.
+// import EmployeeList from '../EmployeeList'
+// import AddEmployee from '../AddEmployee/AddEmployee';
+import './AdminDashboard.css';
 
 type Props= {
   isAdminLoggedIn: boolean
