+++ client/components/admin/DeleteEmployee/DeleteEmployee.css
@@ -1,0 +1,5 @@
+.admin_validation h3 {
+  border: 3px solid rgb(7, 67, 7);
+  color: rgb(7, 67, 7);
+  padding: 10px;
+}
