+++ client/components/admin/UpdateEmployee/UpdateEmployee.css
@@ -1,0 +1,12 @@
+.update_employee {
+  width: 70%;
+  display: block;
+  border: 1px solid black;
+  margin: 5px;
+}
+
+.admin_validation h3 {
+  border: 3px solid rgb(7, 67, 7);
+  color: rgb(7, 67, 7);
+  padding: 10px;
+}
