+++ client/components/admin/UpdateEmployee/UpdateEmployee.tsx
@@ -4,6 +4,7 @@ import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
+import './UpdateEmployee.css';
 
 interface EmployeeProps {
   list: {
