+++ client/css/styles.css
@@ -33,14 +33,6 @@ img {
   margin: 6px;
 }
 
-.add_employee {
-  display: block;
-  border: 1px solid black;
-  margin: 5px;
-  width: 40%;
-  border-radius: 10px;
-}
-
 .add_form {
   border: 1px solid red;
 }
@@ -48,19 +40,6 @@ img {
 .timesheet-grid {
   border: 3px solid rgb(7, 67, 7);
   margin: 6%;
-}
-
-.update_employee {
-  width: 70%;
-  display: block;
-  border: 1px solid black;
-  margin: 5px;
-}
-
-.admin_validation h3 {
-  border: 3px solid rgb(7, 67, 7);
-  color: rgb(7, 67, 7);
-  padding: 10px;
 }
 
 .btn-container {
@@ -87,10 +66,6 @@ img {
   background-color: rgba(67, 66, 66, 0.806);
   color: white;
   padding: 10px 20px;
-}
-
-.admin-dashboard {
-  border: 3px solid rgb(7, 67, 7);
 }
 
 .app-display {
